requirements.txt
openai
numpy
pandas
scikit-learn
faiss-cpu
flask
requests
matplotlib
